requirements.txt
fastapi
uvicorn
sqlalchemy
feedparser
requests
apscheduler
python-dotenv
pydantic
textblob
